requirements.txt
Flask==2.2.2
python-dotenv==0.19.2

flake8
flask-sqlalchemy==3.0.3
Flask-WTF==1.0.1
Flask-Migrate==3.1.0